requirements.txt
clicks
imutils
numpy
